Talos: Move actions next to components
Following the effort to fold related components togethers,
the actions package, which had a one-to-one relationship
with containers (see previous commit) has been folded.
