+++ talos/src/components/administrativeInterface/Action.js
@@ -1,5 +1,5 @@
-import { createActionTypesOf } from '../utils/helpers';
-import { validatePort, validateAddress } from '../services/service.validate'
+import { createActionTypesOf } from '../../utils/helpers';
+import { validatePort, validateAddress } from '../../services/service.validate'
 
 export const CHANGE_ADMINISTRATIVEINTERFACE_ITEMS = createActionTypesOf("CHANGE_ADMINISTRATIVEINTERFACE_ITEMS")
 export const SET_VALID_ADMINISTRATIVEINTERFACE_ITEM = createActionTypesOf("SET_VALID_ADMINISTRATIVEINTERFACE_ITEM")
