+++ talos/src/components/secretSeed/Container.js
@@ -4,7 +4,7 @@ import get from 'lodash/get';
 import {
   changeSecretSeedItems as onChangeSecretSeedItems,
   setValidStateSecretSeedItem as onSetValidStateSecretSeedItem,
-} from '../../actions/secretSeedActions'
+} from './Action'
 
 const mapStateToProps = (state) => ({
   secretSeed: {
